trill: compute forwarding database
Once SPF tree for all node in node database are computed the next step
will be computing forwarding database based on spf tree  rooted in local
node. In fact it is easy to find next hope to any node based on local
spf tree. In case a node is already present on nodedb but not in spf
tree this mean that this node became unreachable. So it has to be
removed from nodedb

Signed-off-by: Ahmed Amamou <ahmed@gandi.net>
Signed-off-by: Kamel Haddadou <kamel@gandi.net>