trill: compute SPF tree only for localnode and tree root
The old code computed the SPF tree for all possible nodes. when having
a small topology this is not a really big deal. but when using a large
topology things start to be complicated an computing time grows
exponentially.
This patch avoid computing SPF tree for all nodes but computes it only
for current node and the already elected tree root. In this way the
computing time grows linearly.

Signed-off-by: Ahmed Amamou <ahmed@gandi.net>