trill: remove a tree root computing
tree root is computed upon receiving a new lsp message.
For safty as thread are asynchounous it is also computed before
computing spf tree ( the LSP thread may remain stuck and thus we will have
no garantie that the root is correct ).
But the removed root computing is unecessary, in fact trill_publish
belong to the same thread of SPF computing. So it is safe to remove it.

Signed-off-by: Ahmed Amamou <ahmed@gandi.net>