trill: fix distribution tree computation
For the computation of a distribution tree not rooted at the current
rbridge, a pseudonode created by the root rbridge of the computed tree
is treated as if it were the root rbridge: his adjacencies are
considered direct adjacencies to the root rbridge. This result in
miscalculated distances and distribution tree.

Signed-off-by: FranÃ§ois Cachereul <f.cachereul@alphalink.fr>
Signed-off-by: Ahmed Amamou <ahmed@gandi.net>